Refactor orchestration mission console layout
